requirements.txt
PyQt5>=5.15.9
PyYAML>=6.0.1
psutil>=5.9.8

